Changes:
- split core
- add create_all, drop_all, run_sync
- add jdbc (stub, non-functional)
- add import structure
- bugfixes
