+++ databasez/types.py
@@ -1,9 +1,10 @@
 from collections.abc import Sequence
-from typing import Any, Callable, Dict, TypeVar
+from typing import TYPE_CHECKING, Any, Callable, Dict, TypeVar
 
-from databasez.interfaces import Record as RecordInterface
+if TYPE_CHECKING:
+    from databasez import interfaces
 
 DictAny = Dict[str, Any]
 
 BatchCallableResult = TypeVar("BatchCallableResult")
-BatchCallable = Callable[[Sequence[RecordInterface]], BatchCallableResult]
+BatchCallable = Callable[[Sequence["interfaces.Record"]], BatchCallableResult]
